small refactoring before the introduction of redis - removed .venv from pylint dir since it tries to lint all deps - moved actions to separate file - actions now inherit from two mixin classes for database and queue - requirements file is now closed after the test
