+++ app/domain/usecases/queue_actions.py
@@ -1,0 +1,6 @@
+import structlog
+
+
+class QueueActions:
+    def __init__(self, logger: structlog.BoundLogger) -> None:
+        self._logger = logger
